requirements.txt
absl-py==0.15.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.2
flatbuffers==1.12
fonttools==4.47.2
gast==0.3.3
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.4
nltk==3.8.1
oauthlib==3.2.2
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.13.2
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.3.0
tqdm==4.66.1
typing-extensions==3.7.4.3
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.12.1
zipp==3.17.0
